Refactor add all ingredients post request so that throwing an error doesn't update the local pantry data
