+++ src/scripts.js
@@ -1,6 +1,6 @@
 import './styles.css';
 import { fetchAll } from './apiCalls';
-import { addAllIngredients } from './post';
+import { createPostRequests, postAll } from './postRequests';
 import Glide from '@glidejs/glide';
 import RecipeRepository from './classes/RecipeRepository';
 import User from './classes/User';
@@ -88,6 +88,33 @@ const initializeApp = () => {
     })
     .catch((err) => console.error(err));
 };
+
+// ------------------- Post to Server ------------------
+
+function addAllIngredients(recipe, user) {
+  const neededIngredients = user.getMissingIngredientsForRecipe(recipe);
+  const requests = createPostRequests(user, neededIngredients);
+  postAll(requests)
+    .then((data) => {
+      data.forEach((response) => {
+        console.log(response);
+      });
+      store.user.addPantryIngredients(
+        store.currentRecipe,
+        store.ingredientsData
+      );
+      populatePantryDisplay();
+      console.log(
+        'MISSING AFTER',
+        store.user.getMissingIngredientsForRecipe(store.currentRecipe)
+      );
+      console.log('NEW PANTRY: ', store.user.pantry);
+    })
+    .catch((err) => {
+      console.error(err);
+      popupError.style.display = 'block';
+    });
+}
 
 // ------------------- EVENT LISTENERS ------------------
 window.addEventListener('load', initializeApp);
@@ -362,12 +389,11 @@ function buildModal(recipe) {
 
 function addRecipesToPantry(event) {
   if (event.target.id === 'addIngredientsBtn') {
-    console.log("RECIPE", store.currentRecipe)
-    addAllIngredients(store.currentRecipe, store.user);
-
-    store.user.addPantryIngredients(store.currentRecipe, store.ingredientsData);
-
-    console.log('NEW PANTRY: ', store.user.pantry);
+    console.log(
+      'MISSING BEFORE',
+      store.user.getMissingIngredientsForRecipe(store.currentRecipe)
+    );
+    addAllIngredients(store.currentRecipe, store.user, store.ingredientsData);
 
     missingIngredientModal.style.display = 'none';
     addIngredientSuccessPopup.style.display = 'block';
